pyproject.toml
[tool.poetry]
name = "propertyapp"
version = "0.1.0"
description = "API v1"
authors = ["Diddychriss"]
readme = "README.md"
packages = [{include = "propertyapp"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
djangorestframework = "^3.14"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
pytest = "^7.4.4"
pytest-django = "^4.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
